requirements.txt
fastapi
uvicorn
python-multipart
geopy
pytz
timezonefinder
vaderSentiment
transformers
torch
sentence-transformers
scikit-learn
accelerate
beautifulsoup4
requests
pdfminer.six
langdetect
simhash
pandas
nltk
faiss-cpu
spacy
spacy-transformers
PyYAML
